Update package and tests
