+++ compiler-tests/src/test/data/dump/fir/aggregation/ContributingTypes.fir.txt
@@ -1,7 +1,10 @@
 FILE: ContributingTypes.kt
     @R|dev/zacsweers/metro/ContributesTo|(scope = <getClass>(Q|dev/zacsweers/metro/AppScope|)) public abstract interface ContributedInterface : R|kotlin/Any| {
+        @R|kotlin/Deprecated|(message = String(This synthesized declaration should not be used directly), level = Q|kotlin/DeprecationLevel|.R|kotlin/DeprecationLevel.HIDDEN|) public abstract interface $$MetroContribution : R|ContributedInterface| {
+        }
+
     }
-    @R|dev/zacsweers/metro/DependencyGraph|(scope = <getClass>(Q|dev/zacsweers/metro/AppScope|)) public abstract interface ExampleGraph : R|kotlin/Any|, R|metro/hints/ContributedInterface| {
+    @R|dev/zacsweers/metro/DependencyGraph|(scope = <getClass>(Q|dev/zacsweers/metro/AppScope|)) public abstract interface ExampleGraph : R|kotlin/Any|, R|ContributedInterface.$$MetroContribution| {
         @R|kotlin/Deprecated|(message = String(This synthesized declaration should not be used directly), level = Q|kotlin/DeprecationLevel|.R|kotlin/DeprecationLevel.HIDDEN|) public final class $$MetroGraph : R|ExampleGraph| {
             public constructor(): R|ExampleGraph.$$MetroGraph| {
                 super<R|kotlin/Any|>()
@@ -18,9 +21,4 @@ FILE: ContributingTypes.kt
 
         }
 
-    }
-FILE: __GENERATED DECLARATIONS__.kt
-    package metro.hints
-
-    @R|dev/zacsweers/metro/internal/Origin|(value = <getClass>(<getClass>(R|ContributedInterface|))) public abstract interface ContributedInterface : R|ContributedInterface| {
     }
