+++ compiler-tests/src/test/data/dump/fir/aggregation/ContributingTypesDependency.fir.txt
@@ -1,15 +1,13 @@
 Module: lib
 FILE: module_lib_ContributingTypesDependency.kt
     @R|dev/zacsweers/metro/ContributesTo|(scope = <getClass>(Q|dev/zacsweers/metro/AppScope|)) public abstract interface ContributedInterface : R|kotlin/Any| {
-    }
-FILE: __GENERATED DECLARATIONS__.kt
-    package metro.hints
+        @R|kotlin/Deprecated|(message = String(This synthesized declaration should not be used directly), level = Q|kotlin/DeprecationLevel|.R|kotlin/DeprecationLevel.HIDDEN|) public abstract interface $$MetroContribution : R|ContributedInterface| {
+        }
 
-    @R|dev/zacsweers/metro/internal/Origin|(value = <getClass>(<getClass>(R|ContributedInterface|))) public abstract interface ContributedInterface : R|ContributedInterface| {
     }
 Module: main
 FILE: module_main_ContributingTypesDependency.kt
-    @R|dev/zacsweers/metro/DependencyGraph|(scope = <getClass>(Q|dev/zacsweers/metro/AppScope|)) public abstract interface ExampleGraph : R|kotlin/Any|, R|metro/hints/ContributedInterface| {
+    @R|dev/zacsweers/metro/DependencyGraph|(scope = <getClass>(Q|dev/zacsweers/metro/AppScope|)) public abstract interface ExampleGraph : R|kotlin/Any| {
         @R|kotlin/Deprecated|(message = String(This synthesized declaration should not be used directly), level = Q|kotlin/DeprecationLevel|.R|kotlin/DeprecationLevel.HIDDEN|) public final class $$MetroGraph : R|ExampleGraph| {
             public constructor(): R|ExampleGraph.$$MetroGraph| {
                 super<R|kotlin/Any|>()
