+++ compiler-tests/src/test/kotlin/dev/zacsweers/metro/compiler/RuntimeTestServices.kt
@@ -1,6 +1,6 @@
 // Copyright (C) 2025 Zac Sweers
 // SPDX-License-Identifier: Apache-2.0
-package dev.zacsweers.metro
+package dev.zacsweers.metro.compiler
 
 import java.io.File
 import org.jetbrains.kotlin.cli.jvm.config.addJvmClasspathRoot
